add marvel api client for comicSeries.
