+++ sample/build.gradle
@@ -72,7 +72,7 @@ dependencies {
   compile 'com.squareup.picasso:picasso:2.5.2'
   compile 'com.karumi:dividers:1.0.3'
   compile 'com.victor:lib:1.0.1' // https://github.com/yankai-victor/Loading
-  compile 'com.karumi:marvelapiclient:0.0.1'
+  compile 'com.karumi:marvelapiclient:0.0.3'
   provided 'com.squareup.dagger:dagger-compiler:1.2.2'
   compile project(':rosie')
 }
