+++ sample/src/main/java/com/karumi/rosie/sample/comics/view/fragment/ComicSeriesFragment.java
@@ -82,7 +82,6 @@ public class ComicSeriesFragment extends RosieFragment implements ComicsSeriesPr
   }
 
   @Override public void showComicSeries(List<ComicSeriesViewModel> comicSeries) {
-    comicSeriesAdapter.clear();
     comicSeriesAdapter.addAll(comicSeries);
     comicSeriesAdapter.notifyDataSetChanged();
     comicSeriesView.setVisibility(View.VISIBLE);
